refactor(effect): split into observability/scope/schema/runtime subfolders
src/rules/effect/ had 22 production children after the v0.0.9 batch,
tripping no-large-folder (max 10). Split the new rules across four
sub-families based on Effect concern:

- effect/observability/  log + span rules (6 files)
  no-console-in-effect, logger-config-at-boot, prefer-annotate-logs,
  require-span-on-exported-effect, handler-requires-span,
  annotate-without-span
- effect/scope/          scoped-resource lifecycle (4 files)
  acquire-release-requires-scope, finalizer-requires-scope,
  fork-requires-lifecycle, runpromise-requires-scoped
- effect/schema/         decoder/boundary rules (3 files)
  no-schema-type-cast, prefer-decode-effect-at-boundary,
  parse-into-schema-requires-effect
- effect/runtime/        Promise interop / concurrency / platform / config (4 files)
  no-promise-all-in-effect, effect-foreach-requires-concurrency,
  prefer-effect-platform, prefer-config-redacted

Top-level effect/ keeps the original 5 (effect-error-erasure,
effect-promise, either-discriminant, no-effect-error-coalescing,
tag-discriminant) plus 4 subfolders = 9 direct children, under the
limit. Each subfolder ships an index.ts that exports its rule map and
a README naming what belongs there.

rule-correctness fixture: prefer-decode-effect-at-boundary now lists
parse-into-schema-requires-effect as a co-fire (same JSON.parse
seed triggers both).

All 909 tests pass. Lint clean (0 errors, 0 warnings).
