+++ src/rules/effect/observability/handler-requires-span.test.ts
