+++ src/rules/effect/runtime/no-promise-all-in-effect.test.ts
