+++ src/rules/effect/runtime/prefer-config-redacted.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from "@typescript-eslint/utils";
 import { AST_NODE_TYPES } from "@typescript-eslint/utils";
-import { createRule } from "../../utils/create-rule.js";
+import { createRule } from "../../../utils/create-rule.js";
 
 const SECRET_NAME_TERMS = [
   /api[_-]?key/i,
