+++ src/rules/effect/schema/no-schema-type-cast.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from "@typescript-eslint/utils";
 import { AST_NODE_TYPES } from "@typescript-eslint/utils";
-import { createRule } from "../../utils/create-rule.js";
+import { createRule } from "../../../utils/create-rule.js";
 
 const SCHEMA_TYPE_PATHS: readonly (readonly string[])[] = [
   ["Schema", "Schema", "Type"],
