+++ src/rules/effect/schema/parse-into-schema-requires-effect.ts
@@ -1,6 +1,6 @@
 import type { TSESTree } from "@typescript-eslint/utils";
 import { AST_NODE_TYPES } from "@typescript-eslint/utils";
-import { createRule } from "../../utils/create-rule.js";
+import { createRule } from "../../../utils/create-rule.js";
 
 const SCHEMA_DECODE_METHODS = new Set([
   "decodeUnknownSync",
