+++ src/rules/effect/scope/finalizer-requires-scope.test.ts
