+++ src/rules/effect/scope/fork-requires-lifecycle.test.ts
